init_dog

+++ 0x0E-structures_typedef/1-init_dog.c
@@ -8,7 +8,7 @@
  *
  */
 
-void init_dog(struct dog *d, char *name, int age, char *owner)
+void init_dog(struct dog *d, char *name, float age, char *owner)
 {
 	if (d)
 	{
